Implemented GUI and User controls for runtime fractal exploration.

+++ include/Zoom.hpp
@@ -10,9 +10,9 @@
  * @brief Represents a single zoom action or position on the fractal image.
  */
 struct Zoom {
-	int x{0};
-	int y{0};
-	double scale{ 0.0 };
+    double x{0.0};
+    double y{0.0};
+    double scale{ 0.0 };
 
 	/**
      * @brief Construct a new Zoom object.
@@ -20,6 +20,6 @@ struct Zoom {
      * @param y The y-coordinate of the zoom center.
      * @param scale The zoom scale factor.
      */
-	Zoom(int x, int y, double scale) : x(x), y(y), scale(scale) {};
+    Zoom(double x, double y, double scale) : x(x), y(y), scale(scale) {};
 };
